telegram-admin/requirements.txt
aiogram==3.0.0b7
PyYAML~=6.0
requests